Split jackson-based features into separate files
